// src/app/shared/interfaces/client.interface.ts
// Generated by https://quicktype.io

export interface ClientResponse {
  id:         number;
  firstName:  string;
  lastName:   string;
  fullName:   string;
  documentId: string;
  document:   Document;
}

export interface Document {
  id:         string;
  name:       string;
  validFrom:  string;
  validUntil: string;
}


export interface ClientInsertResponse {
  client:   Client;
  document: Document;
}

export interface Client {
  firstName: string;
  lastName:  string;
}