// src/app/shared/interfaces/request-type.interface.ts
// Generated by https://quicktype.io

export interface RequestTypeResponse {
  id: number;
  name: string;
  minAmount: number;
  maxAmount: number;
  term: string;
}

export interface ResquestInsertResponse {
  requestTypeId: number,
  description: string,
  plazo: number,
  startDate: string,
  amount: number,
  endDate: string,
  clientId?: number,
  documentId: string
}

export interface RequestTypeCreate {
  name: string,
  minAmount: number
  maxAmount: number
  term: string,
}